pyproject.toml
[project]
name = "doremus_mcp"
version = "0.1.0"
description = "DOREMUS Knowledge Graph MCP Server and Assistant"
authors = [{name = "Simone Fassio", email = "simone.fassio@eurecom.fr"},{name = "Dario Gosmar", email = "dario.gosmar@eurecom.fr"}]
dependencies = [
    "fastapi",
    "requests",
    "langchain",
    "langchain_openai",
    "langchain-mcp-adapters",
    "pandas",
    "matplotlib",
    "pytest",
    "pytest-xdist",
    "jupyter",
    "langgraph-cli[inmem]",
    "python-dotenv",
    "pyppeteer",
    "html2text",
    "rich",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"rdf_assistant" = "src/rdf_assistant"

[tool.setuptools.package-data]
"*" = ["py.typed"]